Move crds_gossip simulator to integration tests
